erase won't mess so much up, heap allocator

+++ fht_ht.hpp
@@ -4,7 +4,6 @@
 #include <assert.h>
 #include <stdint.h>
 #include <sys/mman.h>
-#include <iostream>
 #include <string>
 #include <type_traits>
 
@@ -108,24 +107,23 @@ const uint32_t FHT_HASH_SEED = 0;
 #include <smmintrin.h>
 
 
-static const int8_t VALID_MASK   = ((int8_t)0x80);
-static const int8_t INVALID_MASK = ((int8_t)0x00);
-static const int8_t ERASED_MASK  = ((int8_t)0x01);
+static const int8_t INVALID_MASK = ((int8_t)0x80);
+static const int8_t ERASED_MASK  = ((int8_t)0xC0);
 static const int8_t CONTENT_MASK = ((int8_t)0x7F);
 #define CONTENT_BITS 7
 
 #define FHT_IS_ERASED(tag)  (((tag)) == ERASED_MASK)
 #define FHT_SET_ERASED(tag) ((tag) = ERASED_MASK)
 
-#define FHT_MM_SET(X)     _mm_set1_epi8(X)
-#define FHT_MM_MASK(X, Y) ((uint32_t)_mm_movemask_epi8(_mm_cmpeq_epi8(X, Y)))
-#define FHT_MM_EMPTY(X)                                                        \
-    ((uint32_t)((_mm_movemask_epi8(_mm_cmpeq_epi8(X, FHT_INVALID_VEC)))))
-#define FHT_MM_EMPTY_OR_DEL(X) ((uint32_t)((~_mm_movemask_epi8(X)) & 0xffff))
+#define FHT_SET_INVALID(tag) ((tag) = INVALID_MASK)
 
-static const __m256i FHT_REHASH_VEC  = _mm256_set1_epi8(ERASED_MASK);
-static const __m256i FHT_RESET_VEC   = _mm256_set1_epi8(INVALID_MASK);
-static const __m128i FHT_INVALID_VEC = _mm_set1_epi8(INVALID_MASK);
+#define FHT_MM_SET(X)          _mm_set1_epi8(X)
+#define FHT_MM_MASK(X, Y)      ((uint32_t)_mm_movemask_epi8(_mm_cmpeq_epi8(X, Y)))
+#define FHT_MM_EMPTY(X)        ((uint32_t)_mm_movemask_epi8(_mm_sign_epi8(X, X)))
+#define FHT_MM_EMPTY_OR_DEL(X) ((uint32_t)_mm_movemask_epi8(X))
+
+
+static const __m256i FHT_RESET_VEC = _mm256_set1_epi8(INVALID_MASK);
 
 
 static const uint32_t FHT_MM_LINE      = FHT_TAGS_PER_CLINE / sizeof(__m128i);
@@ -148,7 +146,7 @@ static const uint32_t FHT_MM_IDX_MASK = FHT_MM_IDX_MULT - 1;
       FHT_TO_MASK(tbl_log)) /                                                  \
      FHT_TAGS_PER_CLINE)
 
-#define FHT_GEN_TAG(hash_val) (((hash_val)&CONTENT_MASK) | VALID_MASK)
+#define FHT_GEN_TAG(hash_val) ((hash_val)&CONTENT_MASK)
 #define FHT_GEN_START_IDX(hash_val)                                            \
     (const uint32_t)((hash_val) >> (8 * sizeof(hash_type_t) - 2))
 //////////////////////////////////////////////////////////////////////
@@ -170,6 +168,9 @@ struct DEFAULT_HASH_64;
 // remaps each time table resizes
 template<typename K, typename V>
 struct DEFAULT_MMAP_ALLOC;
+
+template<typename K, typename V>
+struct DEFAULT_ALLOC;
 
 // this will perform significantly better if the copy time on either keys or
 // values is large (it tries to avoid at least a portion of the copying step in
@@ -303,6 +304,11 @@ struct fht_chunk {
     inline void constexpr __attribute__((always_inline))
     delete_tag_n(const uint32_t n) {
         FHT_SET_ERASED(((int8_t * const)this->tags_vec)[n]);
+    }
+
+    inline void constexpr __attribute__((always_inline))
+    invalidate_tag_n(const uint32_t n) {
+        FHT_SET_INVALID(((int8_t * const)this->tags_vec)[n]);
     }
 
     // this undeletes
@@ -659,7 +665,7 @@ struct fht_iterator_t {
         // initialization of new iterator (not from find but from begin) so that
         // it starts at a valid slot
         while (((uint64_t)init_tag_pos) < end &&
-               (!((*init_tag_pos) & VALID_MASK))) {
+               (*init_tag_pos) & INVALID_MASK) {
             if (__builtin_expect(
                     ((uint64_t)(init_tag_pos) % FHT_TAGS_PER_CLINE) ==
                         (FHT_TAGS_PER_CLINE - 1),
@@ -687,7 +693,7 @@ struct fht_iterator_t {
                 this->cur_tag += (sizeof(fht_chunk<K, V>) - FHT_TAGS_PER_CLINE);
             }
             this->cur_tag++;
-        } while (!((*(this->cur_tag)) & VALID_MASK));
+        } while ((*(this->cur_tag)) & INVALID_MASK);
         return *this;
     }
 
@@ -713,7 +719,7 @@ struct fht_iterator_t {
                 this->cur_tag -= sizeof(typename fht_chunk<K, V>::node_t);
             }
             this->cur_tag--;
-        } while (!((*(this->cur_tag)) & VALID_MASK));
+        } while ((*(this->cur_tag)) & INVALID_MASK);
         return *this;
     }
 
@@ -802,7 +808,7 @@ struct fht_iterator_t {
 template<typename K,
          typename V,
          typename Hasher    = DEFAULT_HASH_64<K>,
-         typename Allocator = INPLACE_MMAP_ALLOC<K, V>>
+         typename Allocator = DEFAULT_ALLOC<K, V>>
 struct fht_table {
 
 
@@ -834,6 +840,7 @@ struct fht_table {
     typedef _fht_iterator_t<K, V> fht_iterator;
     //////////////////////////////////////////////////////////////////////
     fht_table(const uint64_t init_size) {
+
         // ensure init_size is above min
         const uint64_t _init_size = init_size > FHT_DEFAULT_INIT_SIZE
                                         ? roundup_next_p2(init_size)
@@ -846,8 +853,16 @@ struct fht_table {
         this->chunks =
             this->alloc_mmap.allocate((_init_size / FHT_TAGS_PER_CLINE));
 
-        this->log_incr = _log_init_size;
-        this->npairs   = 0;
+        // might be faster with _m512 but this is a mile from any critical path
+        // and makes less portable
+        for (uint32_t i = 0; i < (_init_size / FHT_TAGS_PER_CLINE); i++) {
+            for (uint32_t j = 0; j < FHT_MM_LINE; j++) {
+                ((__m256i * const)(this->chunks + i))[0] = FHT_RESET_VEC;
+                ((__m256i * const)(this->chunks + i))[1] = FHT_RESET_VEC;
+            }
+            this->log_incr = _log_init_size;
+            this->npairs   = 0;
+        }
     }
     fht_table() : fht_table(FHT_DEFAULT_INIT_SIZE) {}
 
@@ -910,6 +925,7 @@ struct fht_table {
         fht_chunk<K, V> * const new_chunks =
             this->alloc_mmap.allocate(_num_chunks);
 
+
         uint32_t to_move = 0;
         uint32_t new_starts;
         uint32_t old_start_good_slots;
@@ -930,15 +946,22 @@ struct fht_table {
             __m256i * const set_tags_vec =
                 (__m256i * const)(old_chunk->tags_vec);
 
+            //            __m256i * const set_tags_vec_new =
+            //                (__m256i * const)(new_chunk->tags_vec);
+
 
             // turn all deleted tags -> INVALID (reset basically)
-            set_tags_vec[0] = _mm256_min_epi8(set_tags_vec[0], FHT_REHASH_VEC);
-            set_tags_vec[1] = _mm256_min_epi8(set_tags_vec[1], FHT_REHASH_VEC);
+            set_tags_vec[0] = _mm256_min_epu8(set_tags_vec[0], FHT_RESET_VEC);
+            set_tags_vec[1] = _mm256_min_epu8(set_tags_vec[1], FHT_RESET_VEC);
+
+            //            set_tags_vec_new[0] = FHT_RESET_VEC;
+            //            set_tags_vec_new[1] = FHT_RESET_VEC;
+
             uint64_t j_idx,
-                iter_mask =
-                    ((((uint64_t)_mm256_movemask_epi8(set_tags_vec[1])) << 32) |
-                     (((uint32_t)_mm256_movemask_epi8(set_tags_vec[0])) &
-                      (0xffffffffU)));
+                iter_mask = ~(
+                    (((uint64_t)_mm256_movemask_epi8(set_tags_vec[1])) << 32) |
+                    (((uint32_t)_mm256_movemask_epi8(set_tags_vec[0])) &
+                     (0xffffffffU)));
 
             while (iter_mask) {
                 __asm__("tzcnt %1, %0" : "=r"((j_idx)) : "rm"((iter_mask)));
@@ -980,6 +1003,7 @@ struct fht_table {
                                 std::move(*(old_chunk->get_val_n_ptr(
                                     (const uint32_t)j_idx))));
 
+
                             new_starts += ((1u) << (8 * outer_idx));
                             break;
                         }
@@ -996,6 +1020,14 @@ struct fht_table {
                     else {
                         old_start_good_slots += ((1u) << (8 * start_idx));
                     }
+                }
+            }
+
+            for (uint32_t j = 0; j < FHT_MM_LINE; j++) {
+                const uint32_t inner_idx = (new_starts >> (8 * j)) & 0xff;
+                for (uint32_t _j = inner_idx; _j < FHT_MM_IDX_MULT; _j++) {
+                    new_chunk->set_tag_n(j * FHT_MM_IDX_MULT + _j,
+                                         INVALID_MASK);
                 }
             }
 
@@ -1112,6 +1144,7 @@ struct fht_table {
         // set this while its definetly still in cache
         this->chunks = new_chunks;
 
+
         // iterate through all chunks and re-place nodes
         for (uint32_t i = 0; i < _num_chunks; i++) {
             uint64_t slot_idx = 0;
@@ -1125,10 +1158,10 @@ struct fht_table {
 
             const uint32_t temp_taken_slots =
                 (old_chunk->get_empty_or_del(1) << 16) |
-                (old_chunk->get_empty_or_del(0));
+                (old_chunk->get_empty_or_del(0) & 0xffff);
 
             taken_slots = (old_chunk->get_empty_or_del(3) << 16) |
-                          (old_chunk->get_empty_or_del(2));
+                          (old_chunk->get_empty_or_del(2) & 0xffff);
 
 
             taken_slots = (taken_slots << 32) | temp_taken_slots;
@@ -1137,6 +1170,7 @@ struct fht_table {
             while (taken_slots) {
                 __asm__("tzcnt %1, %0" : "=r"((j_idx)) : "rm"((taken_slots)));
                 taken_slots ^= ((1UL) << j_idx);
+
 
                 const hash_type_t raw_slot =
                     this->hash(old_chunk->get_key_n((const uint32_t)j_idx));
@@ -1159,11 +1193,9 @@ struct fht_table {
                         const uint32_t true_idx =
                             FHT_MM_IDX_MULT * outer_idx + inner_idx;
 
-
                         new_chunk->set_tag_n(
                             true_idx,
                             old_chunk->get_tag_n((const uint32_t)j_idx));
-
                         NEW(K,
                             *(new_chunk->get_key_n_ptr(true_idx)),
                             std::move(*(old_chunk->get_key_n_ptr(
@@ -1173,13 +1205,29 @@ struct fht_table {
                             std::move(*(old_chunk->get_val_n_ptr(
                                 (const uint32_t)j_idx))));
 
+
                         slot_idx += ((1UL) << (8 * outer_idx + 32 * nth_bit));
                         break;
                     }
                 }
             }
+            // set remaining to INVALID_MASK
+            for (uint32_t j = 0; j < FHT_MM_LINE; j++) {
+                const uint32_t inner_idx = (slot_idx >> (8 * j)) & 0xff;
+                for (uint32_t _j = inner_idx; _j < FHT_MM_IDX_MULT; _j++) {
+                    new_chunks[i].set_tag_n(FHT_MM_IDX_MULT * j + _j,
+                                            INVALID_MASK);
+                }
+            }
+            for (uint32_t j = 0; j < FHT_MM_LINE; j++) {
+                const uint32_t inner_idx = (slot_idx >> (8 * j + 32)) & 0xff;
+                for (uint32_t _j = inner_idx; _j < FHT_MM_IDX_MULT; _j++) {
+                    new_chunks[i | _num_chunks].set_tag_n(
+                        FHT_MM_IDX_MULT * j + _j,
+                        INVALID_MASK);
+                }
+            }
         }
-
         // deallocate old table
         this->alloc_mmap.deallocate(
             (fht_chunk<K, V> * const)old_chunks,
@@ -1278,13 +1326,14 @@ struct fht_table {
         uint32_t idx, slot_mask, del_idx = FHT_TAGS_PER_CLINE;
         for (uint32_t j = 0; j < FHT_MM_LINE; j++) {
             const uint32_t outer_idx = (j + start_idx) & FHT_MM_LINE_MASK;
+
             slot_mask = FHT_MM_MASK(tag_match, chunk->tags_vec[outer_idx]);
             while (slot_mask) {
                 __asm__("tzcnt %1, %0" : "=r"((idx)) : "rm"((slot_mask)));
                 const uint32_t true_idx = FHT_MM_IDX_MULT * outer_idx + idx;
+
                 if (__builtin_expect((chunk->compare_key_n(true_idx, new_key)),
                                      1)) {
-
                     return (const int8_t * const)(
                         (((uint64_t)chunk->get_val_n_ptr(true_idx)) |
                          ((1UL) << 48)));
@@ -1328,7 +1377,6 @@ struct fht_table {
                 }
             }
             else if (chunk->get_empty(outer_idx)) {
-
                 chunk->set_key_val_tag(del_idx,
                                        tag,
                                        new_key,
@@ -1364,6 +1412,7 @@ struct fht_table {
                                            tag,
                                            new_key,
                                            std::forward<Args>(args)...);
+
 
                 return ((const int8_t * const)new_chunk) + true_idx;
             }
@@ -1441,7 +1490,6 @@ struct fht_table {
         const __m128i * const tags = (const __m128i * const)(
             (this->chunks + FHT_HASH_TO_IDX(raw_slot, _log_incr)));
 
-
         // by setting valid here we can remove delete check
         const __m128i  tag_match = FHT_MM_SET(FHT_GEN_TAG(raw_slot));
         const uint32_t start_idx = FHT_GEN_START_IDX(raw_slot);
@@ -1464,9 +1512,7 @@ struct fht_table {
 
             // consider adding incrmenter to compiler knows its not infinite
             while (slot_mask) {
-
                 __asm__("tzcnt %1, %0" : "=r"((idx)) : "rm"((slot_mask)));
-
                 const uint32_t true_idx = FHT_MM_IDX_MULT * outer_idx + idx;
 
                 if ((nodes[true_idx].key == key)) {
@@ -1575,7 +1621,12 @@ struct fht_table {
                 __asm__("tzcnt %1, %0" : "=r"((idx)) : "rm"((slot_mask)));
                 const uint32_t true_idx = FHT_MM_IDX_MULT * outer_idx + idx;
                 if ((chunk->compare_key_n(true_idx, key))) {
-                    chunk->delete_tag_n(true_idx);
+                    if (__builtin_expect(chunk->get_empty(outer_idx), 1)) {
+                        chunk->invalidate_tag_n(true_idx);
+                    }
+                    else {
+                        chunk->delete_tag_n(true_idx);
+                    }
                     this->npairs--;
                     return FHT_ERASED;
                 }
@@ -1627,7 +1678,12 @@ struct fht_table {
                 __asm__("tzcnt %1, %0" : "=r"((idx)) : "rm"((slot_mask)));
                 const uint32_t true_idx = FHT_MM_IDX_MULT * outer_idx + idx;
                 if ((nodes[true_idx].key == key)) {
-                    FHT_SET_ERASED(tags8[true_idx]);
+                    if (__builtin_expect(FHT_MM_EMPTY(tags[outer_idx]), 1)) {
+                        FHT_SET_INVALID(tags8[true_idx]);
+                    }
+                    else {
+                        FHT_SET_ERASED(tags8[true_idx]);
+                    }
                     this->npairs--;
                     return FHT_ERASED;
                 }
@@ -1642,12 +1698,10 @@ struct fht_table {
         return FHT_NOT_ERASED;
     }
 
-
     inline constexpr uint64_t
     erase(fht_iterator fht_it) {
         return erase(fht_it->first);
     }
-
 
     void
     clear() {
@@ -1980,11 +2034,6 @@ struct INPLACE_MMAP_ALLOC {
                           0));
             this->cur_size = 2 * this->cur_size;
         }
-        // handling null terms for iterator
-        ((fht_chunk<K, V> *)(this->base_address + old_start_offset))
-            ->set_tag_n(0, 0);
-        ((fht_chunk<K, V> *)(this->base_address + this->start_offset))
-            ->set_tag_n(0, VALID_MASK);
         return (fht_chunk<K, V> *)(this->base_address + old_start_offset);
     }
 
@@ -2000,16 +2049,32 @@ struct DEFAULT_MMAP_ALLOC {
 
     fht_chunk<K, V> *
     allocate(const size_t size) const {
-        fht_chunk<K, V> * const ret = (fht_chunk<K, V> * const)mymmap_alloc(
+        return (fht_chunk<K, V> * const)mymmap_alloc(
             NULL,
             size * sizeof(fht_chunk<K, V>) +
                 1);  // + 1 is in a sense null term for iterator
-        ret[size].set_tag_n(0, VALID_MASK);
-        return ret;
     }
     void
     deallocate(fht_chunk<K, V> * const ptr, const size_t size) const {
-        myMunmap((void * const)ptr, size * sizeof(fht_chunk<K, V>) + 1);
+        myMunmap((void * const)ptr, size * sizeof(fht_chunk<K, V>));
+    }
+};
+
+
+template<typename K, typename V>
+struct DEFAULT_ALLOC {
+    fht_chunk<K, V> *
+    allocate(const size_t size) const {
+        int8_t * const ret = (int8_t * const)aligned_alloc(
+            L1_CACHE_LINE_SIZE,
+            size * sizeof(fht_chunk<K, V>) +
+                1);  // + 1 is in a sense null term for iterator
+        ret[size * sizeof(fht_chunk<K, V>)] = 0;
+        return (fht_chunk<K, V> *)ret;
+    }
+    void
+    deallocate(fht_chunk<K, V> * const ptr, const size_t size) const {
+        free(ptr);
     }
 };
